fix: converted to maven project
