+++ .gitignore
@@ -1,4 +1,1 @@
-*.class
-*.iml
-.idea
-.out
+target
