+++ src/test/java/com/diprosoren15/app/AppTest.java
@@ -1,0 +1,19 @@
+package com.diprosoren15.app;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue() {
+        assertTrue( true );
+    }
+}
